desktop-app/requirements.txt
PySide6>=6.7.0
google-genai>=1.0.0
markdown>=3.6
pyinstaller>=6.0
